+++ src/Hellion.World/Structures/WorldObject.cs
@@ -22,6 +22,8 @@ namespace Hellion.World.Structures
 
         public float Angle { get; set; }
 
+        public float AngleFly { get; set; }
+
         public Vector3 Position { get; set; }
 
         public ICollection<WorldObject> SpawnedObjects { get; set; }